+++ src/main/java/bc1/gream/domain/order/service/SellNowService.java
@@ -1,12 +1,13 @@
 package bc1.gream.domain.order.service;
 
+import bc1.gream.domain.buy.service.BuyService;
 import bc1.gream.domain.gifticon.service.GifticonService;
 import bc1.gream.domain.order.dto.request.SellNowRequestDto;
 import bc1.gream.domain.order.dto.response.SellNowResponseDto;
-import bc1.gream.domain.order.entity.Buy;
+import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.order.entity.Order;
 import bc1.gream.domain.order.mapper.OrderMapper;
-import bc1.gream.domain.order.repository.BuyRepository;
+import bc1.gream.domain.buy.repository.BuyRepository;
 import bc1.gream.domain.order.service.command.OrderCommandService;
 import bc1.gream.domain.user.entity.Coupon;
 import bc1.gream.domain.user.entity.User;
