+++ src/test/java/bc1/gream/domain/order/service/BuyServiceTest.java
@@ -6,14 +6,15 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
+import bc1.gream.domain.buy.service.BuyService;
 import bc1.gream.domain.gifticon.repository.GifticonRepository;
-import bc1.gream.domain.order.dto.request.BuyBidRequestDto;
-import bc1.gream.domain.order.dto.request.BuyNowRequestDto;
-import bc1.gream.domain.order.dto.response.BuyBidResponseDto;
-import bc1.gream.domain.order.dto.response.BuyNowResponseDto;
-import bc1.gream.domain.order.entity.Buy;
+import bc1.gream.domain.buy.dto.request.BuyBidRequestDto;
+import bc1.gream.domain.buy.dto.request.BuyNowRequestDto;
+import bc1.gream.domain.buy.dto.response.BuyBidResponseDto;
+import bc1.gream.domain.buy.dto.response.BuyNowResponseDto;
+import bc1.gream.domain.buy.entity.Buy;
 import bc1.gream.domain.order.entity.Order;
-import bc1.gream.domain.order.repository.BuyRepository;
+import bc1.gream.domain.buy.repository.BuyRepository;
 import bc1.gream.domain.order.repository.OrderRepository;
 import bc1.gream.domain.order.repository.SellRepository;
 import bc1.gream.domain.product.service.query.ProductService;
